% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{associate_discovered_resource}
\alias{associate_discovered_resource}
\title{Associates a discovered resource ID from Application Discovery Service (ADS) with a migration task}
\usage{
associate_discovered_resource(ProgressUpdateStream, MigrationTaskName,
  DiscoveredResource, DryRun = NULL)
}
\arguments{
\item{ProgressUpdateStream}{The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{The identifier given to the MigrationTask.}

\item{DiscoveredResource}{Object representing a Resource.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.}
}
\description{
Associates a discovered resource ID from Application Discovery Service (ADS) with a migration task.
}
\examples{

}
