% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{describe_migration_task}
\alias{describe_migration_task}
\title{Retrieves a list of all attributes associated with a specific migration task}
\usage{
describe_migration_task(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string"
)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] The identifier given to the MigrationTask.}
}
\description{
Retrieves a list of all attributes associated with a specific migration task.
}
\examples{

}
