% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{import_migration_task}
\alias{import_migration_task}
\title{Registers a new migration task which represents a server, database, etc}
\usage{
import_migration_task(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] Unique identifier that references the migration task.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.}
}
\description{
Registers a new migration task which represents a server, database, etc., being migrated to AWS by a migration tool.
}
\details{
This API is a prerequisite to calling the \code{NotifyMigrationTaskState} API as the migration tool must first register the migration task with Migration Hub.
}
\examples{

}
