% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{list_created_artifacts}
\alias{list_created_artifacts}
\title{Lists the created artifacts attached to a given migration task in an update stream}
\usage{
list_created_artifacts(ProgressUpdateStream, MigrationTaskName,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{ProgressUpdateStream}{The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{Unique identifier that references the migration task.}

\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{NextToken}.}

\item{MaxResults}{Maximum number of results to be returned per page.}
}
\description{
Lists the created artifacts attached to a given migration task in an update stream. This API has the following traits:
}
\details{
\itemize{
\item Gets the list of the created artifacts while migration is taking place.
\item Shows the artifacts created by the migration tool that was associated by the \code{AssociateCreatedArtifact} API.
\item Lists created artifacts in a paginated interface.
}
}
\examples{

}
