% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{list_discovered_resources}
\alias{list_discovered_resources}
\title{Lists discovered resources associated with the given MigrationTask}
\usage{
list_discovered_resources(ProgressUpdateStream, MigrationTaskName,
  NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] The name of the MigrationTask.}

\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{NextToken}.}

\item{MaxResults}{The maximum number of results returned per page.}
}
\description{
Lists discovered resources associated with the given \code{MigrationTask}.
}
\section{Accepted Parameters}{
\preformatted{list_discovered_resources(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

