% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{list_migration_tasks}
\alias{list_migration_tasks}
\title{Lists all, or filtered by resource name, migration tasks associated with the user account making this call}
\usage{
list_migration_tasks(NextToken = NULL, MaxResults = NULL,
  ResourceName = NULL)
}
\arguments{
\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{NextToken}.}

\item{MaxResults}{Value to specify how many results are returned per page.}

\item{ResourceName}{Filter migration tasks by discovered resource name.}
}
\description{
Lists all, or filtered by resource name, migration tasks associated with the user account making this call. This API has the following traits:
}
\details{
\itemize{
\item Can show a summary list of the most recent migration tasks.
\item Can show a summary list of migration tasks associated with a given discovered resource.
\item Lists migration tasks in a paginated interface.
}
}
\section{Accepted Parameters}{
\preformatted{list_migration_tasks(
  NextToken = "string",
  MaxResults = 123,
  ResourceName = "string"
)
}
}

