% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{list_progress_update_streams}
\alias{list_progress_update_streams}
\title{Lists progress update streams associated with the user account making this call}
\usage{
list_progress_update_streams(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in \code{NextToken}.}

\item{MaxResults}{Filter to limit the maximum number of results to list per page.}
}
\description{
Lists progress update streams associated with the user account making this call.
}
\section{Accepted Parameters}{
\preformatted{list_progress_update_streams(
  NextToken = "string",
  MaxResults = 123
)
}
}

