% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{notify_application_state}
\alias{notify_application_state}
\title{Sets the migration state of an application}
\usage{
notify_application_state(
  ApplicationId = "string",
  Status = "NOT_STARTED"|"IN_PROGRESS"|"COMPLETED",
  DryRun = TRUE|FALSE
)
}
\arguments{
\item{ApplicationId}{[required] The configurationId in ADS that uniquely identifies the grouped application.}

\item{Status}{[required] Status of the application - Not Started, In-Progress, Complete.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.}
}
\description{
Sets the migration state of an application. For a given application identified by the value passed to \code{ApplicationId}, its status is set or updated by passing one of three values to \code{Status}: \code{NOT_STARTED | IN_PROGRESS | COMPLETED}.
}
\examples{

}
