% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{notify_migration_task_state}
\alias{notify_migration_task_state}
\title{Notifies Migration Hub of the current status, progress, or other detail regarding a migration task}
\usage{
notify_migration_task_state(ProgressUpdateStream, MigrationTaskName, Task,
  UpdateDateTime, NextUpdateSeconds, DryRun = NULL)
}
\arguments{
\item{ProgressUpdateStream}{The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{Unique identifier that references the migration task.}

\item{Task}{Information about the task's progress and status.}

\item{UpdateDateTime}{The timestamp when the task was gathered.}

\item{NextUpdateSeconds}{Number of seconds after the UpdateDateTime within which the Migration Hub can expect an update. If Migration Hub does not receive an update within the specified interval, then the migration task will be considered stale.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.}
}
\description{
Notifies Migration Hub of the current status, progress, or other detail regarding a migration task. This API has the following traits:
}
\details{
\itemize{
\item Migration tools will call the \code{NotifyMigrationTaskState} API to share the latest progress and status.
\item \code{MigrationTaskName} is used for addressing updates to the correct target.
\item \code{ProgressUpdateStream} is used for access control and to provide a namespace for each migration tool.
}
}
\examples{

}
