% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.migrationhub_operations.R
\name{describe_application_state}
\alias{describe_application_state}
\title{Gets the migration status of an application}
\usage{
describe_application_state(
  ApplicationId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The configurationId in ADS that uniquely identifies the grouped application.}
}
\description{
Gets the migration status of an application.
}
\examples{

}
